requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
aiohttp==3.8.1
azure-eventhub
azure-identity
python-dotenv
